+++ packages/segmented-control/src/typing.ts
@@ -1,0 +1,1 @@
+export type IDType = string | number;