+++ packages/typography/src/colors.ts
@@ -11,6 +11,13 @@ export const colors = [
     'secondary-inverted',
     'link',
     'negative',
+    'static-primary-light',
+    'static-secondary-light',
+    'static-tertiary-light',
+    'static-primary-dark',
+    'static-secondary-dark',
+    'static-tertiary-dark',
+    'static-accent',
 ] as const;
 
 export type Color = typeof colors[number];