+++ packages/base-modal/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 5.1.1
+
+### Patch Changes
+
+### [#540](https://github.com/core-ds/core-components/pull/540)
+
+-   Изменена нода с overflow: auto в SidePanelMobile
+
+## 5.1.0
+
+### Minor Changes
+
+### [#494](https://github.com/core-ds/core-components/pull/494)
+
+-   В ModalContext у base-modal добавлен ref на div-обертку модальных окон
+-   У компонентов Modal, SidePanel и BottomSheet обновлён компонент заголовка и изменены основные отступы<br />
+
 ## 5.0.10
 
 ### Patch Changes