+++ packages/confirmation/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 12.4.10
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@7.1.1
+
+## 12.4.9
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@3.2.0
+
 ## 12.4.8
 
 ### Patch Changes