+++ packages/input-autocomplete/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 9.3.11
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@7.1.1
+    -   select@13.2.14
+    -   input@11.1.8
+
+## 9.3.10
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@13.2.13
+
 ## 9.3.9
 
 ### Patch Changes