+++ packages/password-input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.1.9
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-button@6.0.8
+    -   input@11.1.8
+
 ## 4.1.8
 
 ### Patch Changes