+++ packages/select/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 13.2.14
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   base-modal@5.1.1
+    -   bottom-sheet@6.0.1
+    -   modal@7.0.1
+    -   button@7.1.1
+    -   input@11.1.8
+
+## 13.2.13
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   bottom-sheet@6.0.0
+    -   modal@7.0.0
+    -   base-modal@5.1.0
+
 ## 13.2.12
 
 ### Patch Changes