+++ packages/toast/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 4.2.17
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   toast-plate@5.1.14
+
 ## 4.2.16
 
 ### Patch Changes